demo/worker/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
joblib==1.2.0
multidict==6.0.4
numpy==1.24.2
pandas==2.0.0
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2023.3
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
tagoio-sdk==4.0.10
threadpoolctl==3.1.0
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
yarl==1.8.2
